letterboxd-scraper/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6
beautifulsoup4==4.12.2
selenium==4.15.2
